Move MainService class to manager
